Move log_file command tests from test.rs to log_upload.rs
