+++ crates/extensions/c8y_mapper_ext/src/tests.rs
@@ -27,10 +27,8 @@ use tedge_actors::Sender;
 use tedge_actors::SimpleMessageBox;
 use tedge_actors::SimpleMessageBoxBuilder;
 use tedge_actors::WrappedInput;
-use tedge_api::mqtt_topics::Channel;
 use tedge_api::mqtt_topics::EntityTopicId;
 use tedge_api::mqtt_topics::MqttSchema;
-use tedge_api::mqtt_topics::OperationType;
 use tedge_api::CommandStatus;
 use tedge_api::SoftwareUpdateCommand;
 use tedge_file_system_ext::FsWatchEvent;
@@ -41,7 +39,6 @@ use tedge_mqtt_ext::Topic;
 use tedge_test_utils::fs::with_exec_permission;
 use tedge_test_utils::fs::TempTedgeDir;
 use tedge_timer_ext::Timeout;
-use tedge_uploader_ext::UploadResponse;
 
 const TEST_TIMEOUT_MS: Duration = Duration::from_millis(5000);
 
@@ -2216,483 +2213,6 @@ async fn inventory_registers_unknown_entity_once() {
             .any(|message| message == service_register_message),
         "duplicate registration message"
     );
-}
-
-#[tokio::test]
-async fn mapper_converts_smartrest_logfile_req_to_log_upload_cmd_for_main_device() {
-    let cfg_dir = TempTedgeDir::new();
-    let (mqtt, _http, _fs, _timer, _ul, _dl) = spawn_c8y_mapper_actor(&cfg_dir, true).await;
-    let mut mqtt = mqtt.with_timeout(TEST_TIMEOUT_MS);
-
-    skip_init_messages(&mut mqtt).await;
-
-    // Simulate c8y_LogfileRequest SmartREST request
-    mqtt.send(MqttMessage::new(
-        &C8yTopic::downstream_topic(),
-        "522,test-device,logfileA,2013-06-22T17:03:14.123+02:00,2013-06-23T18:03:14.123+02:00,ERROR,1000",
-    ))
-        .await
-        .expect("Send failed");
-
-    let (topic, received_json) = mqtt
-        .recv()
-        .await
-        .map(|msg| {
-            (
-                msg.topic,
-                serde_json::from_str::<serde_json::Value>(msg.payload.as_str().expect("UTF8"))
-                    .expect("JSON"),
-            )
-        })
-        .unwrap();
-
-    let mqtt_schema = MqttSchema::default();
-    let (entity, channel) = mqtt_schema.entity_channel_of(&topic).unwrap();
-    assert_eq!(entity, "device/main//");
-
-    if let Channel::Command {
-        operation: OperationType::LogUpload,
-        cmd_id,
-    } = channel
-    {
-        // Validate the topic name
-        assert_eq!(
-            topic.name,
-            format!("te/device/main///cmd/log_upload/{cmd_id}")
-        );
-
-        // Validate the payload JSON
-        let expected_json = json!({
-            "status": "init",
-            "tedgeUrl": format!("http://localhost:8888/tedge/file-transfer/test-device/log_upload/logfileA-{cmd_id}"),
-            "type": "logfileA",
-            "dateFrom": "2013-06-22T17:03:14.123+02:00",
-            "dateTo": "2013-06-23T18:03:14.123+02:00",
-            "searchText": "ERROR",
-            "lines": 1000
-        });
-
-        assert_json_diff::assert_json_include!(actual: received_json, expected: expected_json);
-    } else {
-        panic!("Unexpected response on channel: {:?}", topic)
-    }
-}
-
-#[tokio::test]
-async fn mapper_converts_smartrest_logfile_req_to_log_upload_cmd_for_child_device() {
-    let cfg_dir = TempTedgeDir::new();
-    let (mqtt, _http, _fs, _timer, _ul, _dl) = spawn_c8y_mapper_actor(&cfg_dir, true).await;
-    let mut mqtt = mqtt.with_timeout(TEST_TIMEOUT_MS);
-
-    skip_init_messages(&mut mqtt).await;
-
-    // Simulate log_upload cmd metadata message
-    mqtt.send(MqttMessage::new(
-        &Topic::new_unchecked("te/device/DeviceSerial///cmd/log_upload"),
-        r#"{"types" : [ "typeA", "typeB", "typeC" ]}"#,
-    ))
-    .await
-    .expect("Send failed");
-
-    mqtt.skip(3).await; //Skip entity registration, mapping and supported log types messages
-
-    // Simulate c8y_LogfileRequest SmartREST request
-    mqtt.send(MqttMessage::new(
-        &C8yTopic::downstream_topic(),
-        "522,test-device:device:DeviceSerial,logfileA,2013-06-22T17:03:14.123+02:00,2013-06-23T18:03:14.123+02:00,ERROR,1000",
-    ))
-    .await
-    .expect("Send failed");
-
-    let (topic, received_json) = mqtt
-        .recv()
-        .await
-        .map(|msg| {
-            (
-                msg.topic,
-                serde_json::from_str::<serde_json::Value>(msg.payload.as_str().expect("UTF8"))
-                    .expect("JSON"),
-            )
-        })
-        .unwrap();
-
-    let mqtt_schema = MqttSchema::default();
-    let (entity, channel) = mqtt_schema.entity_channel_of(&topic).unwrap();
-    assert_eq!(entity, "device/DeviceSerial//");
-
-    if let Channel::Command {
-        operation: OperationType::LogUpload,
-        cmd_id,
-    } = channel
-    {
-        // Validate the topic name
-        assert_eq!(
-            topic.name,
-            format!("te/device/DeviceSerial///cmd/log_upload/{cmd_id}")
-        );
-
-        // Validate the payload JSON
-        let expected_json = json!({
-            "status": "init",
-            "tedgeUrl": format!("http://localhost:8888/tedge/file-transfer/test-device:device:DeviceSerial/log_upload/logfileA-{cmd_id}"),
-            "type": "logfileA",
-            "dateFrom": "2013-06-22T17:03:14.123+02:00",
-            "dateTo": "2013-06-23T18:03:14.123+02:00",
-            "searchText": "ERROR",
-            "lines": 1000
-        });
-
-        assert_json_diff::assert_json_include!(actual: received_json, expected: expected_json);
-    } else {
-        panic!("Unexpected response on channel: {:?}", topic)
-    }
-}
-
-#[tokio::test]
-async fn mapper_converts_log_upload_cmd_to_supported_op_and_types_for_main_device() {
-    let ttd = TempTedgeDir::new();
-    let (mqtt, _http, _fs, _timer, _ul, _dl) = spawn_c8y_mapper_actor(&ttd, true).await;
-    let mut mqtt = mqtt.with_timeout(TEST_TIMEOUT_MS);
-
-    skip_init_messages(&mut mqtt).await;
-
-    // Simulate log_upload cmd metadata message
-    mqtt.send(MqttMessage::new(
-        &Topic::new_unchecked("te/device/main///cmd/log_upload"),
-        r#"{"types" : [ "typeA", "typeB", "typeC" ]}"#,
-    ))
-    .await
-    .expect("Send failed");
-
-    assert_received_contains_str(&mut mqtt, [("c8y/s/us", "118,typeA,typeB,typeC")]).await;
-
-    // Validate if the supported operation file is created
-    assert!(ttd
-        .path()
-        .join("operations/c8y/c8y_LogfileRequest")
-        .exists());
-}
-
-#[tokio::test]
-async fn mapper_converts_log_upload_cmd_to_supported_op_and_types_for_child_device() {
-    let ttd = TempTedgeDir::new();
-    let (mqtt, _http, _fs, _timer, _ul, _dl) = spawn_c8y_mapper_actor(&ttd, true).await;
-    let mut mqtt = mqtt.with_timeout(TEST_TIMEOUT_MS);
-
-    skip_init_messages(&mut mqtt).await;
-
-    // Simulate log_upload cmd metadata message
-    mqtt.send(MqttMessage::new(
-        &Topic::new_unchecked("te/device/child1///cmd/log_upload"),
-        r#"{"types" : [ "typeA", "typeB", "typeC" ]}"#,
-    ))
-    .await
-    .expect("Send failed");
-
-    // Expect auto-registration message
-    assert_received_includes_json(
-        &mut mqtt,
-        [(
-            "te/device/child1//",
-            json!({"@type":"child-device","@id":"test-device:device:child1"}),
-        )],
-    )
-    .await;
-
-    assert_received_contains_str(
-        &mut mqtt,
-        [(
-            "c8y/s/us",
-            "101,test-device:device:child1,child1,thin-edge.io-child",
-        )],
-    )
-    .await;
-    assert_received_contains_str(
-        &mut mqtt,
-        [(
-            "c8y/s/us/test-device:device:child1",
-            "118,typeA,typeB,typeC",
-        )],
-    )
-    .await;
-
-    // Validate if the supported operation file is created
-    assert!(ttd
-        .path()
-        .join("operations/c8y/test-device:device:child1/c8y_LogfileRequest")
-        .exists());
-}
-
-#[tokio::test]
-async fn handle_log_upload_executing_and_failed_cmd_for_main_device() {
-    let cfg_dir = TempTedgeDir::new();
-    let (mqtt, _http, _fs, _timer, _ul, _dl) = spawn_c8y_mapper_actor(&cfg_dir, true).await;
-
-    let mut mqtt = mqtt.with_timeout(TEST_TIMEOUT_MS);
-    skip_init_messages(&mut mqtt).await;
-
-    // Simulate log_upload command with "executing" state
-    mqtt.send(MqttMessage::new(
-        &Topic::new_unchecked("te/device/main///cmd/log_upload/c8y-mapper-1234"),
-        json!({
-            "status": "executing",
-            "tedgeUrl": format!("http://localhost:8888/tedge/file-transfer/test-device/log_upload/typeA-c8y-mapper-1234"),
-            "type": "typeA",
-            "dateFrom": "2013-06-22T17:03:14.123+02:00",
-            "dateTo": "2013-06-23T18:03:14.123+02:00",
-            "searchText": "ERROR",
-            "lines": 1000
-        })
-            .to_string(),
-    ))
-        .await
-        .expect("Send failed");
-
-    // Expect `501` smartrest message on `c8y/s/us`.
-    assert_received_contains_str(&mut mqtt, [("c8y/s/us", "501,c8y_LogfileRequest")]).await;
-
-    // Simulate log_upload command with "failed" state
-    mqtt.send(MqttMessage::new(
-        &Topic::new_unchecked("te/device/main///cmd/log_upload/c8y-mapper-1234"),
-        json!({
-            "status": "failed",
-            "tedgeUrl": format!("http://localhost:8888/tedge/file-transfer/test-device/log_upload/typeA-c8y-mapper-1234"),
-            "type": "typeA",
-            "dateFrom": "2013-06-22T17:03:14.123+02:00",
-            "dateTo": "2013-06-23T18:03:14.123+02:00",
-            "searchText": "ERROR",
-            "lines": 1000,
-            "reason": "Something went wrong"
-        })
-            .to_string(),
-    ))
-        .await
-        .expect("Send failed");
-
-    // Expect `502` smartrest message on `c8y/s/us`.
-    assert_received_contains_str(
-        &mut mqtt,
-        [(
-            "c8y/s/us",
-            "502,c8y_LogfileRequest,\"Something went wrong\"",
-        )],
-    )
-    .await;
-}
-
-#[tokio::test]
-async fn handle_log_upload_executing_and_failed_cmd_for_child_device() {
-    let cfg_dir = TempTedgeDir::new();
-    let (mqtt, _http, _fs, _timer, _ul, _dl) = spawn_c8y_mapper_actor(&cfg_dir, true).await;
-
-    let mut mqtt = mqtt.with_timeout(TEST_TIMEOUT_MS);
-    skip_init_messages(&mut mqtt).await;
-
-    // Simulate log_upload command with "executing" state
-    mqtt.send(MqttMessage::new(
-        &Topic::new_unchecked("te/device/child1///cmd/log_upload/c8y-mapper-1234"),
-        json!({
-            "status": "executing",
-            "tedgeUrl": format!("http://localhost:8888/tedge/file-transfer/child1/log_upload/typeA-c8y-mapper-1234"),
-            "type": "typeA",
-            "dateFrom": "2013-06-22T17:03:14.123+02:00",
-            "dateTo": "2013-06-23T18:03:14.123+02:00",
-            "searchText": "ERROR",
-            "lines": 1000
-        })
-            .to_string(),
-    ))
-        .await
-        .expect("Send failed");
-
-    // Expect auto-registration message
-    assert_received_includes_json(
-        &mut mqtt,
-        [(
-            "te/device/child1//",
-            json!({"@type":"child-device","@id":"test-device:device:child1"}),
-        )],
-    )
-    .await;
-
-    assert_received_contains_str(
-        &mut mqtt,
-        [(
-            "c8y/s/us",
-            "101,test-device:device:child1,child1,thin-edge.io-child",
-        )],
-    )
-    .await;
-
-    // Expect `501` smartrest message on `c8y/s/us/child1`.
-    assert_received_contains_str(
-        &mut mqtt,
-        [(
-            "c8y/s/us/test-device:device:child1",
-            "501,c8y_LogfileRequest",
-        )],
-    )
-    .await;
-
-    // Simulate log_upload command with "failed" state
-    mqtt.send(MqttMessage::new(
-        &Topic::new_unchecked("te/device/child1///cmd/log_upload/c8y-mapper-1234"),
-        json!({
-            "status": "failed",
-            "tedgeUrl": format!("http://localhost:8888/tedge/file-transfer/child1/log_upload/typeA-c8y-mapper-1234"),
-            "type": "typeA",
-            "dateFrom": "2013-06-22T17:03:14.123+02:00",
-            "dateTo": "2013-06-23T18:03:14.123+02:00",
-            "searchText": "ERROR",
-            "lines": 1000,
-            "reason": "Something went wrong"
-        })
-            .to_string(),
-    ))
-        .await
-        .expect("Send failed");
-
-    // Expect `502` smartrest message on `c8y/s/us/child1`.
-    assert_received_contains_str(
-        &mut mqtt,
-        [(
-            "c8y/s/us/test-device:device:child1",
-            "502,c8y_LogfileRequest,\"Something went wrong\"",
-        )],
-    )
-    .await;
-}
-
-#[tokio::test]
-async fn handle_log_upload_successful_cmd_for_main_device() {
-    let ttd = TempTedgeDir::new();
-    let (mqtt, http, _fs, _timer, ul, _dl) = spawn_c8y_mapper_actor(&ttd, true).await;
-    spawn_dummy_c8y_http_proxy(http);
-
-    let mut mqtt = mqtt.with_timeout(TEST_TIMEOUT_MS);
-    let mut ul = ul.with_timeout(TEST_TIMEOUT_MS);
-    skip_init_messages(&mut mqtt).await;
-
-    // Simulate a log file is uploaded to the file transfer repository
-    let uploaded_file = ttd
-        .dir("file-transfer")
-        .dir("test-device")
-        .dir("log_upload")
-        .file("typeA-c8y-mapper-1234");
-
-    // Simulate log_upload command with "successful" state
-    mqtt.send(MqttMessage::new(
-        &Topic::new_unchecked("te/device/main///cmd/log_upload/c8y-mapper-1234"),
-        json!({
-            "status": "successful",
-            "tedgeUrl": "http://localhost:8888/tedge/file-transfer/test-device/log_upload/typeA-c8y-mapper-1234",
-            "type": "typeA",
-            "dateFrom": "2013-06-22T17:03:14.123+02:00",
-            "dateTo": "2013-06-23T18:03:14.123+02:00",
-            "searchText": "ERROR",
-            "lines": 1000
-        })
-        .to_string(),
-    ))
-    .await
-    .expect("Send failed");
-
-    // Uploader gets a upload request and assert that
-    let request = ul.recv().await.expect("timeout");
-    assert_eq!(request.0, "c8y-mapper-1234"); // Command ID
-    assert_eq!(
-        request.1.url,
-        "http://127.0.0.1:8001/c8y/event/events/dummy-event-id-1234/binaries"
-    );
-    assert_eq!(request.1.file_path, uploaded_file.utf8_path());
-
-    // Simulate Uploader returns a result
-    ul.send((
-        request.0,
-        Ok(UploadResponse {
-            url: request.1.url,
-            file_path: request.1.file_path,
-        }),
-    ))
-    .await
-    .unwrap();
-
-    // Expect `503` smartrest message on `c8y/s/us`.
-    assert_received_contains_str(
-        &mut mqtt,
-        [(
-            "c8y/s/us",
-            "503,c8y_LogfileRequest,https://test.c8y.io/event/events/dummy-event-id-1234/binaries",
-        )],
-    )
-    .await;
-}
-
-#[tokio::test]
-async fn handle_log_upload_successful_cmd_for_child_device() {
-    let ttd = TempTedgeDir::new();
-    let (mqtt, http, _fs, _timer, ul, _dl) = spawn_c8y_mapper_actor(&ttd, true).await;
-    spawn_dummy_c8y_http_proxy(http);
-
-    let mut mqtt = mqtt.with_timeout(TEST_TIMEOUT_MS);
-    let mut ul = ul.with_timeout(TEST_TIMEOUT_MS);
-    skip_init_messages(&mut mqtt).await;
-
-    // Simulate a log file is uploaded to the file transfer repository
-    let uploaded_file = ttd
-        .dir("file-transfer")
-        .dir("test-device:device:child1")
-        .dir("log_upload")
-        .file("typeA-c8y-mapper-1234");
-
-    // Simulate log_upload command with "successful" state
-    mqtt.send(MqttMessage::new(
-        &Topic::new_unchecked("te/device/child1///cmd/log_upload/c8y-mapper-1234"),
-        json!({
-            "status": "successful",
-            "tedgeUrl": "http://localhost:8888/tedge/file-transfer/child1/log_upload/typeA-c8y-mapper-1234",
-            "type": "typeA",
-            "dateFrom": "2013-06-22T17:03:14.123+02:00",
-            "dateTo": "2013-06-23T18:03:14.123+02:00",
-            "searchText": "ERROR",
-            "lines": 1000
-        })
-        .to_string(),
-    ))
-    .await
-    .expect("Send failed");
-
-    mqtt.skip(2).await; // Skip child device registration messages
-
-    // Uploader gets a upload request and assert that
-    let request = ul.recv().await.expect("timeout");
-    assert_eq!(request.0, "c8y-mapper-1234"); // Command ID
-    assert_eq!(
-        request.1.url,
-        "http://127.0.0.1:8001/c8y/event/events/dummy-event-id-1234/binaries"
-    );
-    assert_eq!(request.1.file_path, uploaded_file.utf8_path());
-
-    // Simulate Uploader returns a result
-    ul.send((
-        request.0,
-        Ok(UploadResponse {
-            url: request.1.url,
-            file_path: request.1.file_path,
-        }),
-    ))
-    .await
-    .unwrap();
-
-    // Expect `503` smartrest message on `c8y/s/us`.
-    assert_received_contains_str(
-        &mut mqtt,
-        [(
-            "c8y/s/us/test-device:device:child1",
-            "503,c8y_LogfileRequest,https://test.c8y.io/event/events/dummy-event-id-1234/binaries",
-        )],
-    )
-    .await;
 }
 
 #[tokio::test]
